refactor: split monolith files and fix double-autosave bug
- Fix double-autosave bug in retposto_service.sync_all() where
  _autosave_sync_contacts was called twice (once by sync_account,
  once by sync_all loop)
- Split imap.py (546 lines) into imap/ package with 4 files:
  helpers, client, sync, __init__
- Split kontakto_service.py (610 lines): extract VCF to
  kontakto_vcf.py, categories to kontakto_category.py as mixins
- Split retposto_service.py (530 lines): extract signature mgmt
  to retposto_signature.py, remove duplicate 'return True' bug
- Split cli/retposto.py (530 lines) with serci extracted to
  retposto_search.py
- All source files now <= 500 lines
- Update test mock paths for imap/ package change
