+++ src/A_lien/service/kontakto_service.py
@@ -1,6 +1,8 @@
-"""KontaktoService — contacts CRUD with FTS5, categories, VCF import/export.
+"""KontaktoService — contacts CRUD with FTS5, search, and serialization.
 
-Extends A-core CRUDService following the A-encik EncikService pattern.
+Extends A-core CRUDService. VCF and category features are in mixins:
+- KontaktoVCFMixin (kontakto_vcf.py)
+- KontaktoCategoryMixin (kontakto_category.py)
 """
 
 from __future__ import annotations
@@ -11,19 +13,21 @@ from typing import Any
 from A.core.service import CRUDService
 
 from A_lien.data.storage import get_db, KONTAKTOJ_FTS_CONFIG
+from A_lien.service.kontakto_category import KontaktoCategoryMixin
+from A_lien.service.kontakto_vcf import KontaktoVCFMixin
 
 _kontakto_service: KontaktoService | None = None
 
 
-class KontaktoService(CRUDService):
+class KontaktoService(CRUDService, KontaktoVCFMixin, KontaktoCategoryMixin):
     """Contacts CRUD with FTS5 search, categories, and VCF import/export.
 
     Features:
     - JSON serialization for multi-value fields (phones, emails, languages)
     - Core FTS5 full-text search (inherited from CRUDService)
-    - Category management (CRUD for kategorioj table)
+    - Category management (via KontaktoCategoryMixin)
     - Duplicate detection via fuzzy matching (inherited search_fuzzy)
-    - VCF import/export
+    - VCF import/export (via KontaktoVCFMixin)
     """
 
     # JSON columns that are arrays
@@ -286,317 +290,8 @@ class KontaktoService(CRUDService):
         row = self.db.execute_one("SELECT COUNT(*) AS cnt FROM kontaktoj")
         return row["cnt"] if row else 0
 
-    # ── Domain: category management ─────────────────────────────────────────
-
-    def list_categories(self) -> list[dict[str, Any]]:
-        """List all categories."""
-        return self.db.execute(
-            "SELECT * FROM kategorioj ORDER BY nomo ASC"
-        )
-
-    def create_category(self, nomo: str, koloro: str = "") -> dict[str, Any]:
-        """Create a new category.
-
-        Args:
-            nomo: Category name (unique)
-            koloro: Optional color code
-
-        Returns:
-            Created category dict
-        """
-        from datetime import datetime, timezone
-        import uuid
-
-        now = datetime.now(timezone.utc).isoformat()
-        data = {
-            "uuid": str(uuid.uuid4()),
-            "nomo": nomo,
-            "koloro": koloro,
-            "kreita_je": now,
-            "modifita_je": now,
-        }
-        with self.db.transaction() as conn:
-            conn.execute(
-                "INSERT INTO kategorioj (uuid, nomo, koloro, kreita_je, modifita_je) "
-                "VALUES (?, ?, ?, ?, ?)",
-                (data["uuid"], nomo, koloro, now, now),
-            )
-        return data
-
-    def update_category(self, uuid: str, data: dict[str, Any]) -> dict[str, Any] | None:
-        """Update a category."""
-        from datetime import datetime, timezone
-
-        data["modifita_je"] = datetime.now(timezone.utc).isoformat()
-        set_clauses = [f"{k} = ?" for k in data.keys()]
-        values = list(data.values()) + [uuid]
-        with self.db.transaction() as conn:
-            conn.execute(
-                f"UPDATE kategorioj SET {', '.join(set_clauses)} WHERE uuid = ?",
-                values,
-            )
-        row = self.db.execute_one("SELECT * FROM kategorioj WHERE uuid = ?", (uuid,))
-        return row
-
-    def delete_category(self, uuid: str) -> bool:
-        """Delete a category.
-
-        Args:
-            uuid: Category UUID
-
-        Returns:
-            True if deleted, False if not found
-        """
-        cursor = self.db.execute_one(
-            "SELECT COUNT(*) AS cnt FROM kategorioj WHERE uuid = ?", (uuid,)
-        )
-        if not cursor or cursor["cnt"] == 0:
-            return False
-        with self.db.transaction() as conn:
-            conn.execute("DELETE FROM kategorioj WHERE uuid = ?", (uuid,))
-        return True
-        return True
-
-    # ── Domain: VCF import/export ───────────────────────────────────────────
-
-    def import_vcf(self, path: str) -> int:
-        """Import contacts from a VCF file.
-
-        Args:
-            path: Path to .vcf file
-
-        Returns:
-            Number of contacts imported
-
-        Raises:
-            ImportError: If vobject library is not installed
-            FileNotFoundError: If VCF file not found
-        """
-        try:
-            import vobject
-        except ImportError:
-            raise ImportError(
-                "vobject library required for VCF import. "
-                "Install: pip install vobject"
-            )
-
-        from pathlib import Path
-
-        path_obj = Path(path)
-        if not path_obj.exists():
-            raise FileNotFoundError(f"VCF file not found: {path}")
-
-        with path_obj.open("r", encoding="utf-8") as f:
-            content = f.read()
-
-        count = 0
-        for vcard in vobject.readComponents(content):
-            contact = self._vcard_to_contact(vcard)
-            if contact and contact.get("plena_nomo"):
-                self.create(contact)
-                count += 1
-
-        return count
-
-    def export_vcf(self, uuid: str | None = None, path: str | None = None) -> str:
-        """Export contact(s) to VCF format.
-
-        Args:
-            uuid: Single contact UUID (None = export all)
-            path: Optional output file path (None = return string)
-
-        Returns:
-            VCF string (if path is None)
-
-        Raises:
-            ImportError: If vobject library is not installed
-        """
-        try:
-            import vobject
-        except ImportError:
-            raise ImportError(
-                "vobject library required for VCF export. "
-                "Install: pip install vobject"
-            )
-
-        contacts = [self.get(uuid)] if uuid else self.list()
-
-        lines: list[str] = []
-        for contact in contacts:
-            if not contact:
-                continue
-            lines.append(self._contact_to_vcard(contact))
-
-        vcf_text = "\n".join(lines)
-
-        if path:
-            from pathlib import Path
-
-            Path(path).write_text(vcf_text, encoding="utf-8")
-
-        return vcf_text
-
-    # ── VCF conversion helpers ──────────────────────────────────────────────
-
-    @staticmethod
-    def _vcard_to_contact(vcard: Any) -> dict[str, Any]:
-        """Convert a vobject vCard to a contact dict."""
-        from datetime import datetime, timezone
-        import uuid
-
-        now = datetime.now(timezone.utc).isoformat()
-        contact: dict[str, Any] = {
-            "uuid": str(uuid.uuid4()),
-            "kreita_je": now,
-            "modifita_je": now,
-            "lingvoj": [],
-            "telefonnumeroj": [],
-            "retposhtadresoj": [],
-            "kampoj": {},
-            "kategorioj": [],
-            "konfirmita": 0,
-        }
-
-        # Name
-        if hasattr(vcard, "n"):
-            given = str(vcard.n.value.given) if vcard.n.value.given else ""
-            family = str(vcard.n.value.family) if vcard.n.value.family else ""
-            contact["nomo"] = given
-            contact["familia_nomo"] = family
-            contact["plena_nomo"] = f"{given} {family}".strip()
-
-        if hasattr(vcard, "fn"):
-            fn_val = str(vcard.fn.value) if vcard.fn.value else ""
-            if not contact.get("plena_nomo"):
-                contact["plena_nomo"] = fn_val
-            if not contact.get("nomo") and not contact.get("familia_nomo"):
-                contact["nomo"] = fn_val
-
-        # Email
-        if hasattr(vcard, "email"):
-            for email in vcard.contents.get("email", []):
-                addr = str(email.value) if email.value else ""
-                if addr:
-                    if not contact.get("retposto"):
-                        contact["retposto"] = addr
-                    contact["retposhtadresoj"].append({
-                        "valoro": addr,
-                        "etikedo": email.params.get("TYPE", [""])[0] if hasattr(email, "params") else "",
-                        "cxefa": not bool(contact.get("retposto")),
-                    })
-
-        # Phone
-        if hasattr(vcard, "tel"):
-            for tel in vcard.contents.get("tel", []):
-                num = str(tel.value) if tel.value else ""
-                if num:
-                    contact["telefonnumeroj"].append({
-                        "valoro": num,
-                        "etikedo": tel.params.get("TYPE", [""])[0] if hasattr(tel, "params") else "",
-                        "cxefa": len(contact["telefonnumeroj"]) == 0,
-                    })
-
-        # Organization
-        if hasattr(vcard, "org"):
-            org_val = str(vcard.org.value) if vcard.org.value else ""
-            if org_val:
-                contact["organizo"] = org_val
-
-        # Categories
-        if hasattr(vcard, "categories"):
-            cats = str(vcard.categories.value) if vcard.categories.value else ""
-            if cats:
-                contact["kategorioj"] = [c.strip() for c in cats.split(",")]
-
-        # Note
-        if hasattr(vcard, "note"):
-            note_val = str(vcard.note.value) if vcard.note.value else ""
-            if note_val:
-                contact["noto"] = note_val
-
-        return contact
-
-    @staticmethod
-    def _contact_to_vcard(contact: dict[str, Any]) -> str:
-        """Convert a contact dict to vCard 3.0 string."""
-        try:
-            import vobject
-        except ImportError:
-            raise ImportError("vobject library required for VCF export")
-
-        card = vobject.vCard()
-
-        # Name (N)
-        card.add("n")
-        family = contact.get("familia_nomo", "")
-        given = contact.get("nomo", "")
-        card.n.value = vobject.vcard.Name(family=family, given=given)
-
-        # Full name (FN)
-        card.add("fn")
-        card.fn.value = contact.get("plena_nomo", "") or (
-            f"{given} {family}".strip()
-        )
-
-        # Email
-        email = contact.get("retposto", "")
-        if email:
-            card.add("email")
-            card.email.value = email
-            card.email.type_param = "INTERNET"
-
-        # Additional emails from JSON array
-        retposhtadresoj = contact.get("retposhtadresoj", [])
-        if isinstance(retposhtadresoj, str):
-            try:
-                retposhtadresoj = json.loads(retposhtadresoj)
-            except (json.JSONDecodeError, TypeError):
-                retposhtadresoj = []
-        for addr in retposhtadresoj:
-            val = addr.get("valoro", "")
-            if val and val != email:
-                card.add("email")
-                card.email.value = val
-                card.email.type_param = "INTERNET"
-
-        # Phone
-        telefonnumeroj = contact.get("telefonnumeroj", [])
-        if isinstance(telefonnumeroj, str):
-            try:
-                telefonnumeroj = json.loads(telefonnumeroj)
-            except (json.JSONDecodeError, TypeError):
-                telefonnumeroj = []
-        for tel in telefonnumeroj:
-            val = tel.get("valoro", "")
-            if val:
-                card.add("tel")
-                card.tel.value = val
-                card.tel.type_param = tel.get("etikedo", "VOICE")
-
-        # Organization
-        org = contact.get("organizo", "")
-        if org:
-            card.add("org")
-            card.org.value = [org]
-
-        # Categories
-        kategorioj = contact.get("kategorioj", [])
-        if isinstance(kategorioj, str):
-            try:
-                kategorioj = json.loads(kategorioj)
-            except (json.JSONDecodeError, TypeError):
-                kategorioj = []
-        if kategorioj:
-            card.add("categories")
-            card.categories.value = ",".join(kategorioj)
-
-        # Note
-        note = contact.get("noto", "")
-        if note:
-            card.add("note")
-            card.note.value = note
-
-        return card.serialize()
+    # ── Domain: category management — provided by KontaktoCategoryMixin ────
+    # ── Domain: VCF import/export — provided by KontaktoVCFMixin ───────────
 
 
 def get_kontakto_service() -> KontaktoService:
